Split utils/index.test.ts into per-file test files
