+++ app/src/lib/game-modes/werewolf/utils/targeting.test.ts
@@ -1,0 +1,55 @@
+import { describe, it, expect } from "vitest";
+import { getTargetablePlayers } from "./targeting";
+
+const players = [
+  { id: "owner", name: "Owner" },
+  { id: "p1", name: "Alice" },
+  { id: "p2", name: "Bob" },
+  { id: "p3", name: "Charlie" },
+];
+
+describe("getTargetablePlayers", () => {
+  it("excludes the game owner", () => {
+    const result = getTargetablePlayers(players, "owner", [], "", null, []);
+    expect(result.map((p) => p.id)).toEqual(["p1", "p2", "p3"]);
+  });
+
+  it("excludes dead players", () => {
+    const result = getTargetablePlayers(players, "owner", ["p2"], "", null, []);
+    expect(result.map((p) => p.id)).toEqual(["p1", "p3"]);
+  });
+
+  it("excludes multiple dead players", () => {
+    const result = getTargetablePlayers(
+      players,
+      "owner",
+      ["p1", "p3"],
+      "",
+      null,
+      [],
+    );
+    expect(result.map((p) => p.id)).toEqual(["p2"]);
+  });
+
+  it("returns all non-owner players when no one is dead", () => {
+    const result = getTargetablePlayers(players, "owner", [], "", null, []);
+    expect(result).toHaveLength(3);
+  });
+
+  it("handles undefined ownerPlayerId", () => {
+    const result = getTargetablePlayers(players, undefined, [], "", null, []);
+    expect(result).toHaveLength(4);
+  });
+
+  it("returns empty array when all non-owner players are dead", () => {
+    const result = getTargetablePlayers(
+      players,
+      "owner",
+      ["p1", "p2", "p3"],
+      "",
+      null,
+      [],
+    );
+    expect(result).toHaveLength(0);
+  });
+});
